requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
torch
torchvision
opencv-python
pillow
tqdm
numpy